renamed some files and folder
