+++ a1_hello/HelloGoodbye.java
@@ -6,7 +6,7 @@
  *  message in reverse order).
  */
 
-package hello;
+package a1_hello;
 import edu.princeton.cs.algs4.StdIn;
 
 public class HelloGoodbye {
