+++ a1_hello/RandomWord.java
@@ -1,4 +1,4 @@
-package hello;
+package a1_hello;
 
 import edu.princeton.cs.algs4.StdIn;
 import edu.princeton.cs.algs4.StdRandom;
