+++ a2_search_algo/Problems/p01_SearchInInfiniteArray.java
@@ -1,4 +1,4 @@
-package Problems;
+package a2_search_algo.Problems;
 
 /**
  * @Date : 11/11/22
