+++ a2_search_algo/Problems/p02_FindCeilingInArray.java
@@ -1,4 +1,4 @@
-package Problems;
+package a2_search_algo.Problems;
 
 /**
  * @Date : 08/11/2022
