+++ a2_search_algo/Problems/p03_PeakElementInBitonicArray.java
@@ -1,4 +1,4 @@
-package Problems;
+package a2_search_algo.Problems;
 
 public class p03_PeakElementInBitonicArray {
     static int PeakElement(int[] sample) {
